Create "origin" remote for GitRepository when it does not exist
The get_git_repository function was breaking when the "origin" remote
was unexistent for a cached repository.

+++ lib/repository.py
@@ -51,13 +51,16 @@ def get_git_repository(remote_repo_url, parent_dir_path):
     if os.path.exists(repo_path):
         MAIN_REMOTE_NAME = "origin"
         repo = GitRepository(repo_path)
-        previous_url = repo.remotes[MAIN_REMOTE_NAME].url
-        if previous_url != remote_repo_url:
-            LOG.debug("Updating {name}'s repository URL from '{previous_url}' "
-                      "to '{new_url}'"
-                      .format(name=name, previous_url=previous_url,
-                              new_url=remote_repo_url))
-            repo.delete_remote(MAIN_REMOTE_NAME)
+        if any(remote.name == MAIN_REMOTE_NAME for remote in repo.remotes):
+            previous_url = repo.remotes[MAIN_REMOTE_NAME].url
+            if previous_url != remote_repo_url:
+                LOG.debug("Removing previous {name}'s repository remote with "
+                          "URL '{previous_url}'"
+                          .format(name=name, previous_url=previous_url))
+                repo.delete_remote(MAIN_REMOTE_NAME)
+        if not any(remote.name == MAIN_REMOTE_NAME for remote in repo.remotes):
+            LOG.debug("Creating {name}'s repository remote with URL '{url}'"
+                      .format(name=name, url=remote_repo_url))
             repo.create_remote(MAIN_REMOTE_NAME, remote_repo_url)
         return repo
     else:
